Added optional "fields" parameter to table export
If not specified, exports all columns. If specified, will only export those columns which are listed.
Columns on table can be specified, but one can also specify columns of the form "x.y" if is a column
with a FK constraint on another table. This will result in column "y" being returned from the referenced
table.